Detects whether a pattern is in a character string

Function that detects if one or more patterns are in a string. The patterns can be chained, by default this is a regex search but special flags be triggered with a specific syntax, supports negation.

str_is(
  x,
  ...,
  fixed = FALSE,
  ignore.case = FALSE,
  word = FALSE,
  or = FALSE,
  pattern = NULL,
  envir = parent.frame()
)

str_which(
  x,
  ...,
  fixed = FALSE,
  ignore.case = FALSE,
  word = FALSE,
  or = FALSE,
  pattern = NULL,
  envir = parent.frame()
)

Arguments

x

A character vector.

...

Character scalars representing the patterns to be found. By default they are (perl) regular-expressions. Use ' & ' or ' | ' to chain patterns and combine their result logically (ex: '[[:alpha:]] & \\d' gets strings containing both letters and numbers). You can negate by adding a ! first (ex: "!sepal$" will return TRUE for strings that do not end with "sepal"). Add flags with the syntax 'flag1, flag2/pattern'. Available flags are: 'fixed', 'ignore', 'word', 'verbatim'. Ex: "ignore/sepal" would get "Sepal.Length" (wouldn't be the case w/t 'ignore'). Shortcut: use the first letters of the flags. Ex: "if/dt[" would get "DT[i = 5]" (flags 'ignore' + 'fixed'). The flag 'verbatim' does not parse logical operations. For 'word', it adds word boundaries to the pattern. See the documentation of this argument.

fixed

Logical scalar, default is FALSE. Whether to trigger a fixed search instead of a regular expression search (default).

ignore.case

Logical scalar, default is FALSE. If TRUE, then case insensitive search is triggered.

word

Logical scalar, default is FALSE. If TRUE then a) word boundaries are added to the pattern, and b) patterns can be chained by separating them with a comma, they are combined with an OR logical operation. Example: if word = TRUE, then pattern = "The, mountain" will select strings containing either the word 'The' or the word 'mountain'.

or

Logical, default is FALSE. In the presence of two or more patterns, whether to combine them with a logical "or" (the default is to combine them with a logical "and").

pattern

(If provided, elements of ... are ignored.) A character vector representing the patterns to be found. By default a (perl) regular-expression search is triggered. Use ' & ' or ' | ' to chain patterns and combine their result logically (ex: '[[:alpha:]] & \\d' gets strings containing both letters and numbers). You can negate by adding a ! first (ex: "!sepal$" will return TRUE for strings that do not end with "sepal"). Add flags with the syntax 'flag1, flag2/pattern'. Available flags are: 'fixed', 'ignore', 'word', 'verbatim'. Ex: "ignore/sepal" would get "Sepal.Length" (wouldn't be the case w/t 'ignore'). Shortcut: use the first letters of the flags. Ex: "if/dt[" would get "DT[i = 5]" (flags 'ignore' + 'fixed'). The flag 'verbatim' does not parse logical operations. For 'word', it adds word boundaries to the pattern. See the documentation of this argument.

Value

It returns a logical vector of the same length as x.

The function str_which returns a numeric vector.

Details

The internal function used to find the patterns is base::grepl() with perl = TRUE.

Functions

str_which(): Returns the indexes of the values in which a pattern is detected

Generic regular expression flags

All stringmagick functions support generic flags in regular-expression patterns. The flags are useful to quickly give extra instructions, similarly to usual regular expression flags.

Here the syntax is "flag1, flag2/pattern". That is: flags are a comma separated list of flag-names separated from the pattern with a slash (/). Example: str_which(c("hello...", "world"), "fixed/.") returns 1. Here the flag "fixed" removes the regular expression meaning of "." which would have otherwise meant "any character". The no-flag verion str_which(c("hello...", "world"), ".") returns 1:2.

Alternatively, and this is recommended, you can collate the initials of the flags instead of using a comma separated list. For example: "if/dt[" will apply the flags "ignore" and "fixed" to the pattern "dt[".

The four flags always available are: "ignore", "fixed", "word" and "magic".

"ignore" instructs to ignore the case. Technically, it adds the perl-flag "(?i)" at the beginning of the pattern.

"fixed" removes the regular expression interpretation, so that the characters ".", "$", "^", "[" (among others) lose their special meaning and are treated for what they are: simple characters.

"word" adds word boundaries ("\\b" in regex language) to the pattern. Further, the comma (",") becomes a word separator. Technically, "word/one, two" is treated as "\b(one|two)\b". Example: str_clean("Am I ambushed?", "wi/am") leads to " I ambushed?" thanks to the flags "ignore" and "word".

"magic" allows to interpolate variables inside the pattern before regex interpretation. For example if letters = "aiou" then str_clean("My great goose!", "magic/[{letters}] => e") leads to "My greet geese!"

See also

Author

Laurent R. Berge

Examples

# NOTA: using `str_get` instead of `str_is` may lead to a faster understanding 
#       of the examples 

x = smagick("/one, two, one... two, microphone, check")

# default is regular expression search
# => 3 character items
str_is(x, "^...$")
#> [1] FALSE

# to trigger fixed search use the flag 'fixed'
str_is(x, "fixed/...")
#> [1] TRUE
# you can just use the first letter
str_is(x, "f/...")
#> [1] TRUE

# to negate, use '!'
str_is(x, "!f/...")
#> [1] FALSE
# or directly in the pattern
str_is(x, "f/!...")
#> [1] FALSE

# you can combine several patterns with "&" or "|"
str_is(x, "one & c")
#> [1] TRUE
str_is(x, "one | c")
#> [1] TRUE

#
# word: adds word boundaries
#

# compare
str_is(x, "one")
#> [1] TRUE
# with
str_is(x, "w/one")
#> [1] TRUE

# words can be chained with commas (it is like an OR logical operation)
str_is(x, "w/one, two")
#> [1] TRUE
# compare with
str_is(x, "w/one & two")
#> [1] TRUE
# remember that you can still negate
str_is(x, "w/one & !two")#' 
#> [1] FALSE

# you can combine the flags
# compare
str_is(x, "w/ONE")
#> [1] FALSE
# with
str_is(x, "wi/ONE")
#> [1] TRUE

#
# str_which
#

# it works exactly the same way as str_is
# Which are the items containing an 'e' and an 'o'?
str_which(x, "e", "o")
#> [1] 1
# equivalently
str_which(x, "e & o")
#> [1] 1