Chains basic operations to character vectors

Simple tool to perform multiple operations to character vectors.

str_ops(x, op, pre_unik = NULL)

Arguments

x

A character vector. If not a character vector but atomistic (i.e. not a list), it will be converted to a character vector.

op

Character scalar. Character scalar containing the comma separated values of operations to perform to the vector. The 50+ operations are detailed in the help page of smagick().

pre_unik

Logical scalar, default is NULL. Whether to first unique the vector before applying the possibly costly string operations, and merging back the result. For very large vectors with repeated values the time gained can be substantial. By default, this is TRUE for vector of length 1M or more.

Value

In general it returns a character vector. It may be of a length different from the original one, depending on the operations performed.

Details

This function is a simple wrapper around smagick. Formally, str_ops(x, "op1, op2") is equivalent to smagick("{op1, op2 ? x}").

See also

Author

Laurent R. Berge

Examples